### 2019-05-04 #298 @potterzot
Submission: rnassqs - access the USDA-NASS 'Quick Stats' data through their API
Submitting Author: Nicholas A. Potter (@potterzot)
Repository: https://github.com/potterzot/rnassqs
Version submitted: 0.4.0
Version accepted: TBD

---

```
Package: rnassqs
Title: Access the NASS 'Quick Stats' API
Version: 0.4.0.9000
  person('Nicholas', 'Potter', email='potter.nicholas@gmail.com', role = c('aut', 'cre')),
  person('Joseph', 'Stachelek', email='', role = c('ctb')),
  person('Julia', 'Piaskowski', email='', role = c('ctb')))
Maintainer: Nicholas Potter <potter.nicholas@gmail.com>
Description: Interface to access data via the United States Department of
  Agricultre's National Agricultural Statistical Service (NASS) 'Quick Stats'
  web API <https://quickstats.nass.usda.gov/api>. Convenience functions
  facilitate building queries based on available parameters and valid parameter
  values.
  httr,
  here,
```

Data retrieval because 'rnassqs' allows access the the NASS 'Quick Stats' API to fetch data.

Target audience is those who want to automate or reproducibly fetch data from 'Quick Stats', including agronomists, economists, and others working with agricultural data. Scientific applications include analysis of agricultural data by administrative region (e.g. county, state, watershed), economic analysis of policies that affect agriculture, and sociological/demographic analysis of agricultural producers over time.

None that I have been able to find.

#297, responded to by @noamross

    - [X] The package is deposited in a long-term repository with the DOI: 10.5281/zenodo.2662520
