### 2019-05-08 #298 @lmullen
@potterzot I will be the editor for this peer review.

Here are my editorial checks.

### Editor checks:

---

#### Editor comments

In running R CMD check I get the following:

```
❯ checking top-level files ... NOTE
  Non-standard files/directories found at top level:
    ‘paper.bib’ ‘paper.md’
```

Please either add those files too `.Rbuildignore` or put them in `inst/`.

In `paper.md` there are a few typos, including these. You might look over the paper one more time.

- "be one of the most difficult stages of research make reproducible" should be "to make"
- "explicitly constructing html GET requests" should be "HTTP GET requests." (HTTP is the protocol that often, but not usually in the case of an API like this one, returns HTML.)

Could you clarify, please, what is tested when the user has not provided an API key? For tests that you are skipping if no API key is available, is it possible to stub or mock the API? The testthat package includes some mocking functions, as well as `testthat::expect_equal_to_reference()` which can be used for that purpose.

I am in the process of looking for peer reviewers.

---

Reviewers:
Due date:
