### 2019-05-09 a3e820c
Address editorial comments in ropensci review.
- Expand and fix errors in paper.md
- add paper.md and paper.bib to .Rbuildignore
- fix misspellings
- remove date from DESCRIPTION
