+++ paper.md
@@ -1,49 +1,116 @@
 ---
-title: "The rnassqs R package: access the USDA National Agricultural Statistics Service (NASS) 'Quick Stats' API"
-tags:
-  - R
-  - reproducibility
-  - high-performance computing
-  - pipeline
-  - workflow
-  - Make
+title: "rnassqs: An `R` package to access agricultural data via the USDA 
+  National Agricultural Statistics Service (USDA-NASS) 'Quick Stats' API"
 authors:
-  - name: Nicholas A. Potter
-    orcid: 0000-0002-3410-3732
-    email: potter.nicholas@gmail.com
-    affiliation: 1
-affiliations:
-  - name: Washington State University
-    index: 1
-date: 01 June 2019
+- affiliation: 1
+  email: potter.nicholas@gmail.com
+  name: Nicholas A. Potter
+  orcid: 0000-0002-3410-3732
+date: "09 June 2019"
+output: pdf_document
 bibliography: paper.bib
+tags:
+- R
+- API
+- reproducibility
+affiliations:
+- index: 1
+  name: Washington State University
 ---
 
 # Summary
 
-The [rnassqs](https://github.com/ropensci/rnassqs) `R` package [@rnassqs] is an API wrapper for the United States Department of Agriculture
-National Agricultural Statistics Service (NASS) ‘[Quick
-Stats](https://quickstats.nass.usda.gov/)’ API [@quickstats]. The core
-functionality allows the user to query and parse agricultural data from
-‘Quick Stats’ in a reproducible and automated way. It manages API
-authentication by setting a system environmental variable for the
-duration of the `R` session. Several convenience functions allow the
-user to easily query common data, as well as see the parameters and
-valid parameter values that are available with which to query. The query
-can return both JSON and CSV formats, and parsing converts both of these
-formats to a data.frame object.
+The [rnassqs](https://github.com/ropensci/rnassqs) `R` package [@rnassqs] is an
+API wrapper for the United States Department of Agriculture
+National Agricultural Statistics Service (USDA-NASS) '[Quick
+Stats](https://quickstats.nass.usda.gov/)' API [@quickstats]. The core
+functionality allows the user to query agricultural data from 'Quick Stats' in 
+a reproducible and automated way. 
 
-Without a programmatic interface to an API, the initial collection of
-data can be one of the most difficult stages of research make
-reproducible. The [rnassqs](https://github.com/ropensci/rnassqs) package
-makes the collection of data reproducible, improving both the
-reliability of the research and the accessibility for others seeking to
-replicate or expand on it.
+`rnassqs` manages API authentication by setting a system environmental variable 
+for the duration of the `R` session. Convenience functions facilitate querying
+common data. Users can also use `rnassqs` to query the list of data parameters 
+and available values for a given parameter (for example, to see the commodities 
+available). The query can request data in JSON or CSV formats, and parses that
+data into a `data.frame` object.
 
-In addition, developers of web tools or generated reports can use
-[rnassqs](https://github.com/ropensci/rnassqs) to repeatedly query data,
-which would have previously required explicitly constructing html GET
-requests.
+`rnassqs` was written to address issues that arise from accessing data via a 
+point-and-click interface, particularly in terms of automation and 
+reproducibility of data requests. 
+
+
+# About USDA-NASS data and 'Quick Stats'
+
+'Quick Stats' is a web interface to access data produced by USDA-NASS. The data 
+comes primarily from the Census of Agriculture [@agCensus], but also includes
+data from USDA-NASS surveys on a wide range of topics. The Census of Agriculture
+is conducted every five years in years ending in '2' and '7'. The most recent
+census was completed in 2017. 'Quick Stats' provides access to census data 
+beginning in 1997, and some survey data as far back as 1850.
+
[…]
