+++ README.Rmd
@@ -99,17 +99,17 @@ Install should be simple via either devtools or CRAN:
 
 ## API Key
 
-To use the NASS Quick Stats API you need an [API key](http://quickstats.nass.usda.gov/api). There are several ways of clueing the rnassqs package in to your api key. You can set the variable explicitly and pass it to functions, like so
+To use the NASS Quick Stats API you need an [API key](http://quickstats.nass.usda.gov/api). There are several ways of setting your API key. You can set the variable explicitly and pass it to functions, like so
 
     params <- list(...)                    # parameters for query 
-    api_key <- "<your api key here>"       # api key
+    api_key <- "<your api key here>"       # API key
     data <- nassqs(params, key = api_key)  # query and return data
 
-Alternatively, you can set the api key as an environmental variable either by adding it to your `.Renviron` like so:
+Alternatively, you can set the API key as an environmental variable either by adding it to your `.Renviron` like so:
 
     NASSQS_TOKEN="<your api key here>"
 
-or by setting it explicitly in the console by calling `rnassqs::nassqs_auth()`. This will prompt you to enter the api key if not set, and return the value of the api key if it is set. If you do not set the key and you are running the session interactively, R will ask you for the key when you try to issue a query.
+or by setting it explicitly in the console by calling `rnassqs::nassqs_auth()`. This will prompt you to enter the API key if not set, and return the value of the API key if it is set. If you do not set the key and you are running the session interactively, R will ask you for the key when you try to issue a query.
 
 ## Usage
 
