### 2019-05-09 #298 @potterzot
@lmullen thank you for your comments. I've made a commit to address your comments.

Regarding this note:

> Could you clarify, please, what is tested when the user has not provided an API key? For tests that you are skipping if no API key is available, is it possible to stub or mock the API? The testthat package includes some mocking functions, as well as testthat::expect_equal_to_reference() which can be used for that purpose.

Tests in [tests/testthat/test-oncran.R](https://github.com/potterzot/rnassqs/blob/master/tests/testthat/test-oncran.R#L55) beginning on line 55 include mock API calls. The tests make the request, specifying that the function return the GET request URL rather than actually make the request, and that request is compared to the correct URL. There are three API paths to test:

- __api_GET__: tested in [`test_that("nass_GET forms a correct URL", ...)`](https://github.com/potterzot/rnassqs/blob/master/tests/testthat/test-oncran.R#L76)
- __get_param_values__: tested in [`test_that("nassqs_param_values forms a correct URL", ...)`](https://github.com/potterzot/rnassqs/blob/master/tests/testthat/test-oncran.R#L70)
- __get_counts__: tested in [`test_that("nassqs_record_county forms a correct URL", ...)`](https://github.com/potterzot/rnassqs/blob/master/tests/testthat/test-oncran.R#L64)

There are additional mock API tests that follow those, but those are for convenience functions for making specific requests, e.g. `nassqs_area` and `nassqs_yield`, which wrap `nassqs_GET`.

Tests in [`tests/testthat/test-local.R`](https://github.com/potterzot/rnassqs/blob/master/tests/testthat/test-local.R) make actual API calls using an API key, and are not possible on CRAN.

Is there a better way of organizing tests that makes it clear where the API mock tests are done and where the actual API call tests are done?
