### 2019-05-28 1f3e5ab
fix license misspelling and paper updates.
