### 2019-06-19 #298 @lmullen
Thanks to our reviewers for agreeing to take on this package.

Reviewer: @adamhsparks
Reviewer: @nealrichardson
Due date: 2019-07-11

You can find the [guide for reviewers here](http://devguide.ropensci.org/reviewerguide.html). Please let me know if you have any questions.
