### 2019-07-13 #298 @adamhsparks
- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be auto-generated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

* Spell check the package, in particular the DESCRIPTION file's Line 12, `Agricultre`

* It's a bit of opsec, but I think the README and vignette should be more explicit about keeping the API keys secret and not embedding them in scripts. Along with this, highlighting why it's useful to use the `.Renviron` file for this purpose would be nice (you don't have to worry about adding it to .gitignore for one). For new users just how to modify the .Renviron is likely to be confusing at best and frustrating at worst. Perhaps it would be nice to include an example that shows how to use `usethis::edit_r_environ()` for this purpose to streamline the illustrations and examples.

* The DESCRIPTION file's description text might need " ' " around "API". I had to do this for the nasapower package to have it accepted on CRAN.

* I always encourage including a proper CITATION file in /inst, see <https://github.com/ropensci/nasapower/blob/master/inst/CITATION> for an example

* The paper.md file has Git conflicts in it that need to be resolved

* There is no statement about contributions or a code-of-conduct present

* I feel a bit like the functionality for the end-user is a bit overly complicated. This sort of package to me should just return data in a data.frame or list or some other R object that I can easily work with. In some cases I get a server response, in others I can ask for raw JSON or other file formats. I don't really see a good reason for this in R. I guess the functionality could be offered, but I think it needs to be hidden under a few layers for more advanced users that might want to use it. Rather it should just be a simple query for the data that I want and returning it in a format that I can readily use in R without extra steps. Just fetch the data and format it for me in one single step. I think it's possible with this package, but the documentation is a bit convoluted and not clear to me.

#### Comments on Vignette

When I try using the `nassqs_token()` I receive and error message.

```r
> library(rnassqs)
> nassqs_token()
Error in nassqs_token() : could not find function "nassqs_token"
```

*********

This example works as expected,

```r
nassqs_fields()
[1] "agg_level_desc"        "asd_code"
 [3] "asd_desc"              "begin_code"
 [5] "class_desc"            "commodity_desc"
 [7] "congr_district_code"   "country_code"
 [9] "country_name"          "county_ansi"
[11] "county_code"           "county_name"
[13] "CV"                    "domaincat_desc"
[15] "domain_desc"           "end_code"
[17] "freq_desc"             "group_desc"
[19] "load_time"             "location_desc"
[21] "prodn_practice_desc"   "reference_period_desc"
[23] "region_desc"           "sector_desc"
[25] "short_desc"            "state_alpha"
[27] "state_ansi"            "state_name"
[29] "state_fips_code"       "statisticcat_desc"
[31] "source_desc"           "unit_desc"
[33] "util_practice_desc"    "Value"
[35] "watershed_code"        "watershed_desc"
[37] "week_ending"           "year"
[39] "zip_5"
```

however,

```r
?nassqs_fields()
```

returns a help file that says,

```
Deprecated: Return list of NASS QS parameters.
Description
Deprecated. Use nassqs_params() instead.
```

Suggest updating vignette to match most recent functionality.

************

Another error occurs with this example from the vignette.

```r
> rnassqs::nassqs_field_values(field = 'unit_desc')
Error: 'nassqs_field_values' is not an exported object from 'namespace:rnassqs'
```

************

The "All together" section script is not functional.
```r
fields <- nassq_fields()
Error in nassq_fields() : could not find function "nassq_fields"
```

#### Comments on functions

* Function names are not consistent with GET being all caps but parse and check not. In the vignette text `nassqs_parse()` is referred to as PARSE in all caps. Consistency in function naming will help the end user.

* It is entirely up to the package authors how to organise the functions, but I find the current structure confusing at best. Typically when I see a file with just the package name, it is just there to provide the help file for the package with author information, references and other basic info. The `nassqs.R` file has several functions in it for the package along with `release_questions()`, which I find to be extremely odd. This function is not something that should be in the package and exposed to end-users.

My suggestion is remove `release_questions()` entirely and split out the functions into their own files with the function name being the file-name. This helps make it easier to keep the functions organised and updated. Splitting the functions out is entirely up to the authors if they wish to implement this structure, however I feel that removing `release_questions()` is necessary.

* Why is `base_URL` given as a parameter that the user can modify? The documentation even says it "probably" should not be changed. I would just hard-code it and not give the user any possibility of changing it. I can't see any good reason for doing this. If the URL changes, then the package should be updated to reflect the changes.

#### Comments on documentation

* When packages are mentioned in the documentation, wrap them in \pkg{httr} for proper formatting to indicate that you are referring to a package. Likewise, R can be written as \R for special formatting.

* The text \code{jsonlite::fromJSON} should be written as \code{\link[jsonlite]{fromJSON}} so that it links to the help file for this function.

* Titles for function help files should be written in title case.

* All exported functions should have examples in the documentation, `nassqs_check()` does not have any examples.

* The example for  `nassqs_param_values()` is commented out making it difficult to follow. Examples should not be commented out and should be clear an easy to follow.

```r
  # See all values available for the statisticcat_desc field. Values may not
  # be available in the context of other parameters you set, for example
  # a given state may not have any 'YIElD' in blueberries if they don't grow
  # blueberries in that state.
  # Requires an API key:
  #nassqs_param_values("statisticcat_desc", key = "my api key")
```

Should appear as

```r
  # See all values available for the statisticcat_desc field. Values may not
  # be available in the context of other parameters you set, for example
  # a given state may not have any 'YIElD' in blueberries if they don't grow
  # blueberries in that state.
  # Requires an API key:

  nassqs_param_values("statisticcat_desc", key = "my api key")
```

* I find the documentation for `nassqs_GET()` confusing. I don't understand the first example.

```r
> params = list(commodity_name="CORN",
+               year=2012,
+               agg_level_desc = "STATE",
+               state_alpha = "WA",
+               statisticcat_desc = "YIELD")
> nassqs_GET(params)
Response [https://quickstats.nass.usda.gov/api/api_GET?key=FB4A48D4-62F3-3904-B1E1-D718FDC5EB76&commodity_name=CORN&year=2012&agg_level_desc=STATE&state_alpha=WA&statisticcat_desc=YIELD&format=JSON]
  Date: 2019-07-13 04:48
  Status: 200
  Content-Type: application/json
  Size: 148 kB
```

What do I do with this response value? How is this response yields for corn in 2012 in Washington?

Does the end-user even need to interface with this function or should it be hidden and used by the other functions in the package that return data in `data.frames` or other R objects?

* `nassqs_params()` lacks examples

* The example for `nassqs_parse()` could be easier to follow.

```r
# Set parameters and make the request
params = list(commodity_name="CORN",
              year=2012,
              agg_level_desc = "STATE",
              state_alpha = "WA",
              statisticcat_desc = "YIELD")
req <- nassqs_GET(params)
nassqs_parse(req, as = "data.frame")
```

would be more clear as

```r
# Set parameters and make the request
params <- list(
  commodity_name = "CORN",
  agg_level_desc = "STATE",
  state_alpha = "WA",
  statisticcat_desc = "YIELD"
)
req <- nassqs_GET(params)
corn <- nassqs_parse(req, as = "data.frame")
head(corn)
```

* `nassqs_parse()` `as` is unclear. It states it indicates the data type returned, but doesn't list the data types aside from in the `usage` section, which indicates that a `list` is possible, but this doesn't appear to be documented? The `@return` section says `a data frame or raw text of the content from the request.`

* Any functions that query an external server for data and may fail or take an extended period of time to run should have the examples wrapped in a \donttest{} for CRAN but still allowing for local testing. I see most examples are wrapped but not all that run an external query.

#### Comments on code style

* In most cases the code is clearly written, in some cases the style is inconsistent with lack of spaces around a "=", _e.g._ Line 73 of nassqs.R. This also applies in the examples for documentation. Also, single and double quotes are used interchangeably. I find it easier to follow if only one style is used in all cases as there are cases where single quotes only may be used and so forth and so on.

* The operator used to assign in the examples also switches between `=` and `<-`. Only one should be used consistently.

* Wrap code at 80 characters for ease of reading and for those of us that don't have editor windows that expand beyond 80 chars wide.
