### 2019-07-15 #298 @nealrichardson
---

Looks like I get to be the infamous "Reviewer 2" this time :)

Since Adam dutifully got his review in on time and I'm delayed a couple of days, I'll say at the start that I agree generally with all of his points (with one minor exception, noted below), and in the interest of brevity, I've tried to avoid reiterating comments he made.

I greatly appreciate packages like this and the effort that goes into making them. This is the kind of package I wish existed when I was doing my dissertation--would have made at least the data retrieval and management part of research a lot cleaner. So, thank you for your contribution.

Style and code-hygiene comments aside, my main suggestion to you is to think about how you can orient the package towards the R user and their needs. The purpose of this package should be to encapsulate your hard-earned knowledge of how this API works and enable data scientists to access that data as naturally as possible. However, the package currently seems to expose an R interface centered around the needs of the API and not the R user. For example, the functions that seem to make up the public interface for the package are mostly stuck in "helpers.R", as if they're ancillary when really they should be front and center.

Here's a more concrete example: reading the vignette discussion about the comparison operators, it looks like the way to get data on corn production in Virginia since 2012 is

```r
nassqs(list(commodity_desc="CORN", year__GE=2012, state_alpha="VA")))
```

But when I think about the R code that I want to type to get that data, it looks more like

```r
nassqs(commodity == "corn", state == "va", year >= 2012)
```

That is, not case-sensitive, handles the different ways commodities and states (for example) could be referenced (integer code, abbreviation, long name), and more naturally handles the comparison operators (can translate `year >= 2012` to `year__GE=2012`). You could go even farther and have a `dplyr` syntax where you can `filter()` rows and `select()` columns, so that the R user is essentially describing the data.frame they want to get, and your package figures out how to turn that into API requests, handle pagination, etc., and return the shape of data that an R user expects.

I'm not saying you need to do exactly this/all of this now (though some things, like case insensitvity, would be easy enough), but more to suggest ideas for the future and to show what I mean in terms of thinking of the R interface and not just what the API demands. When I wrap APIs (and in general write packages), I like to start by thinking about the R code I want (users) to type, and that usually means doing more work in the package to mask the awkwardness of the underlying HTTP API.

Another way to get more R-user-centric would be to improve the documentation, which I found to be thin, often tautological (`@param api_path the API path`), and not helpful for an inexperienced user. Documentation is a bit of work but [it's worth doing](https://enpiar.com/2019/02/17/writing-docs-is-better-than-having-docs/), especially for how it can help focus you on what the intended user needs to know in order to get value from the package--and how you can make what details they need to know as minimal as possible.

A good way to see what the docs will look like to your users is to use `pkgdown` to build the package website. You don't need to publish this (though it's nice if you do), but I've found that even doing it locally makes it really visible to me where my documentation is not the most helpful or beautiful.

## Observations from using the package

I got a token and did the example on the readme (corn yields in Virginia and Pennsylvania for since 2000).

```
> library(rnassqs)
> Sys.setenv(NASSQS_TOKEN="REDACTED")
> params <- list("commodity_desc"="CORN",
+                   "year__GE"=2000,
+                   "state_alpha"=c("VA", "PA"),
+                   "statisticcat_desc"="YIELD")
> df <- nassqs(params)
```

That all worked as expected. That said,

* It was pretty slow. Slow enough that I wondered if the process was hung. The resulting data.frame is `4752 x 39`, so not huge. I'm not sure what was slow, but it might be worth considering how to show progress. I believe `httr` has some tools for showing progress if it's slow because of the network, for example. But maybe network isn't the constraint. Granted, I made one request, so I'm not sure how widespread the issue is, but it was distracting.
* It was not at all obvious to me that I would get 39 columns back. It would help if the documentation perhaps described the result, beyond saying "returns a data frame". And is it possible to request fewer columns? Some are clearly duplicate information (state_name, state_alpha).
* It's odd to me that `year` is numeric but everything else in the data.frame is character, even when there are very obviously numeric columns in there. If I'm going to do anything useful with this crop yield data I just downloaded, the first thing I'm going to have to do is `df$Value <- as.numeric(df$Value)` so it seems worthwhile to return properly formatted data.
* It was surprising that the one column name that the [package code changes](https://github.com/potterzot/rnassqs/blob/master/R/nassqs.R#L178) gets a rather unfriendly name (`"CV (%)"`), while the other columns follow a standard, well-behaved format.

## Specific code notes

* General style
  * be consistent with single vs. double quotes
  * don't need to quote/backtick list names (unless there's a space, which isn't the case anywhere here) e.g. prefer `list(param=param)` over `list("param"=param)`
  * My one point of disagreement with Adam: it's ok to have `nassqs_GET()` capitalized: HTTP request methods are supposed to be capitalized (they're case sensitive, [per the RFC](https://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html)). FWIW [in `httr`](https://httr.r-lib.org/reference/index.html), the request methods are capitalized while other acronyms (JSON etc.) are not.
  * `#' @importFrom httr function_name` rather than `httr::function_name`
  * Check out [markdown formatting in roxygen2](https://cran.r-project.org/web/packages/roxygen2/vignettes/markdown.html): will clean up your `\code{\link{}}`s
  * `if (something) { result }` inline is bad form: either drop `{}` or make multiline. See https://style.tidyverse.org/syntax.html#indenting
* data.R:
  * can do `nassqs_fields <- nassqs_params`, don't need `function () that()`
  * Give documentation for what those params mean? Or link to API docs?
* nassqs.R
  * Agree that `release_questions()` doesn't belong in the package
  * nassqs_parse
    * L127 docs not correct since it doesn't necessarily return a data frame
    * "as" list is different from `nassqs`
    * as "raw" is misleading; it's not `raw` in the R sense, it's "character". or as="text" if you want to be like httr
    * L153: when is not a response object? Also better to use `inherits()` rather than class ==. Also can return() early rather than if/else
    * L159 warning: can't be tested because if condition is met, L156 would have errored. Suggests better test coverage needed
    * L177 read.csv
    * L179: no else case?
  * nassqs_auth
    * IMO this function should just be `function (key) Sys.setenv(NASSQS_TOKEN = key)`. What's the value of the interactive mode? It's reasonable to expect user to set auth at the beginning of the session.
    * Not sure the dual purpose (set and get token) is worth it. In places where you want to read the token, just Sys.getenv()
    * L214 don't need to end lines with ;
  * nassqs (and mostly relevant for helpers.R functions too)
    * Is this the function people should be calling? If so, the file would be more readable if you put this function at the top of the file
    * A nicer interface would accept the "params" as ... rather than require you to wrap them in list() yourself. I don't think you need to pass ... to nassqs_GET (if any of its args are important to expose, add them as named arguments)
    * Related, you can use [`@inheritParams`](http://r-pkgs.had.co.nz/man.html#dry2) to document the function args in only one place and pull them in in the other functions that have the same
  * nassqs_GET
    * L35 docs: make that an href
    * I'd remove `key` from the function signature and just Sys.getenv(). L75-78 is unnecessary complexity
    * I'd remove `base_url` from the signature (since it "should probably never be changed"). If you want it configurable for some reason, allow it to be set by an option and in the function get it like `base_url <- getOption("rnasssqs.base_url", "https://quickstats.nass.usda.gov/api/")`
    * I'd remove `url_only` from the signature. It's confusing to have a function than can return very different things based on an argument. If that's functionality you need, I'd factor out the URL query assembly, ending in `build_url()`, to its own function, and have `rnassqs_GET()` call that and then `httr::GET()` the result. If you only have this argument for unit testing, I have an alternative proposal (see below).
    * It's odd that the function allows "XML" format but `rnassqs_parse` errors if it gets XML.
    * Consider removing "format" from the signature altogether since it can be set in params. Also, why would I care what content-type I'm querying in if my end result is going to be a data.frame?
  * nassqs_check: what do I do if my request is too large? Give some recommendation for how to fix it. If it means that you need to make paginated requests, how would I do that? An even friendlier solution would be for you to handle the pagination for the user so they don't need to know/care about this API constraint.
* helpers.R
  * `expand_list()` is probably the only function there that fits with what I'd expect helpers to be (i.e. internal functions). I'd give it [`@keywords internal`](http://r-pkgs.had.co.nz/man.html#roxygen-comments) so it doesn't list in the help index, and possibly just document it with comments and not formal documentation. (And if you're going to document it, explain what it actually does, which is not obvious.)
* Vignette
  * `.secret` file is used in tests and supporting code, and it's discussed in the vignette, but it's not actually supported in the package (i.e. the nassqs_auth() doesn't look for a .secret file). And IIUC the vignette recommendation is incomplete: you'd still have to set or pass it in even if you read it in with readLines(). IMO I'd just drop that discussion and rely on the environment variable, or say that if you wanted to store it in some file (called whatever), you could read it and set it in your session like `Sys.setenv(NASSQS_TOKEN = readLines(file))`.
  * Don't `rnassqs::` your own functions, not necessary
  * Looks like you should `Suggest` `assertthat` since your vignette uses it.
* Tests
  * oncran
    * You may find the [`httptest` package](https://enpiar.com/r/httptest/) useful for these tests. Rather than using the `url_only` argument (which I suspect you only have for tests), you could use `expect_GET(nassqs(...), url)`
    * `httptest` would also let you supply mock responses so you could test the full request/response/parse flow naturally. What you have with the the .rds testdata files works, of course, but `httptest` could make those tests more readable and maintainable.
    * FYI there is `testthat::skip_if()`, and you could of course define your own `skip_if_interactive()` (but you take my suggestion to drop the interactive behavior of the auth function, then you can just delete these tests)
    * `withr::with_envvar` can help manipulate environment variables within tests (`withr` is depended on by testthat, so it's a "free" dependency, in case that's a concern)
    * L133-151: maybe you want one of these to assert that you can get the "text" content, but mostly you're just asserting that `httr::content()` works.
  * local
    * Again, I'd drop the .secret file (and thus the `here` dependency)
    * L37: `is.numeric(as.numeric(...))` is always true so this test doesn't do anything; also for future reference, `testthat::expect_true` may help in some places
    * L56: `httptest` would let you [test handling of error responses without a network connection](https://enpiar.com/2017/06/21/7-hard-testing-problems-made-easy-by-httptest/)

## Miscellaneous

* Should add `stats` and `utils` to `Imports` in DESCRIPTION
* `R CMD check` shows this NOTE--maybe because the file is huge? Consider (re)moving the link, obfuscating it so that CRAN doesn't try to download it, or something.
```
Found the following (possibly) invalid URLs:
  URL: ftp://ftp.nass.usda.gov/quickstats
    From: README.md
    Status: Error
    Message: libcurl error code 6:
      	Could not resolve host: ftp.nass.usda.gov
```
* Would be good to add rnassqs.Rcheck and rnassqs_*.tar.gz to .Rbuildignore and .gitignore so that artifacts from `R CMD build` and `check` locally don't get checked in
* .Rbuildignore includes a misspelling (paoer.pdf)
* Re: spelling, check out the `spelling` package.
