### 2019-07-16 #298 @lmullen
@adamhsparks and @nealrichardson: Thanks to both of you for these thorough and detailed reviews.

@potterzot Please take a look at these comments from the reviewers. There is lots of good advice in here. It does seem like some of it will require some pretty fundamental reassessment of the package's user-facing interface. Do you think you can make revisions within two weeks, our typical deadline? i.e., by July 29?
