### 2019-07-16 #298 @lmullen
@potterzot Sure. If you can plan on a July 29 deadline, that would be best, but an extension would be fine if it becomes necessary.
