### 2019-07-23 #298 @potterzot
Hi @adamhsparks, thank you for your time and willingness to help. I've finished making the straight-foward changes you suggested, and am thinking about the larger issues, which seem to boil down to two related issues:

1. How much should we expose to the user
2. How can we simplify the interface to make it more usable

The main function of the package is `nassqs()`, which fetches and returns parsed data. But I exposed `nassqs_GET()` and `nassqs_parse()` because I wanted to make it possible for advanced users to deal with any edge cases that might come up. Perhaps that's not really a concern here, and it would make sense to hide both of those functions. Working with output from `nassqs_GET()` requires some knowledge of the `httr` package and how requests work, so it would only be useful to someone who wants to see the raw results and knows what to do with them. `nassqs_parse()` on the other hand doesn't do much more than what `jsonlite::fromJSON` does, so I think it could be hidden without really removing any control from the user. My goal was to make it easy to use, but also to allow a user to dig into the deeper mechanism if necessary. This may be born out of personal frustration from when I've not understood what a hidden function is doing and had to root around in the source code to figure it out.

I propose hiding `nassqs_parse()`, expanding the documentation for `nassqs_GET()` so that it clearly states that the main function is `nassqs()` and that in general that function should not be needed. What do you think about that and about the larger issue of usability?
