### 2019-07-23 #298 @adamhsparks
I think that sounds reasonable. The documentation can always be structured with the meat in the main vignette and then more advanced usage in another or farther down the page.
