### 2019-07-24 7aba631
Major updates based on ropensci reviewer responses
Changes include:
* Many code style standarizations
* Simplification of authentication
* Expanded test coverage and use of httptest::with_mock_api()
* Better clarification in documentation
* Improved README and vignette
* Change in syntax to allow for lower case query parameter values
* Change in syntax to allow for specifying each parameter as a separate
function argument rather than as a single list
