+++ R/nassqs.R
@@ -1,269 +1,18 @@
 #' rnassqs is a wrapper for the United States Department of Agriculture's
 #' National Agricultural Statistical Service (NASS) 'Quick Stats' API to enable 
-#' getting NASS 'Quick Stats' data directly from R.  Based on the httr API 
+#' getting NASS 'Quick Stats' data directly from \R.  Based on the httr API 
 #' package guide.
 #'
 #' The functions in this package facilitate getting data from NASS 'Quick Stats'.
-#' It handles the API key checking and storage, authorization, and fetching of data.
+#' It handles the API key checking and storage, authorization, and fetching of 
+#' data.
 #' @author Nicholas Potter
 #' @name rnassqs-package
 #' @aliases rnassqs
 #' @docType package
-#' @title rnassqs-package: Use the NASS 'Quick Stats' API from R.
+#' @title rnassqs-package: Access the NASS 'Quick Stats' API
 #' @keywords package rnassqs-package
 #' @references \url{http://quickstats.nass.usda.gov}
 #' @seealso \url{http://quickstats.nass.usda.gov/api}
 NULL
 
-#' Questions to ask before release.
-release_questions <- function() {
-  c(
-    "Have you submitted to winbuilder with devtools::check_win_devel?",
-    "Have you submitted to winbuilder with devtools::check_win_release?",
-    "Have you updated the version number?",
-    "Have you run devtools::check() and updated cran-comments.md?",
-    "Have you updated NEWS.md?",
-    "Have you updated README.Rmd?"
-  )
-}
-
-
-#' Issue a GET request to the NASS 'Quick Stats' API
-#' 
-#' This is the core function, which several other rnassqs functions use to 
-#' request data from the NASS 'Quick Stats' API: 
-#' https://quickstats.nass.usda.gov/api. 
-#'
-#' @export
-
-#' @param params a named list of values to be queried.
-#' @param key your API key. If not provided the function will check for an 
-#' environmental variable and if not found, will prompt for your API key.
-#' @param api_path the API path.
-#' @param base_url the base API url. This should probably never be changed.
-#' @param url_only (logical) if TRUE, returns the URL and makes no API call.
-#' @param format format of returned data. JSON by default, but can also be XML 
-#' or CSV. Can also be set as a parameter.
-#' @return data returned in the format specified.
-#' @examples
-#' \donttest{
-#' # Yields for corn in 2012 in Washington
-#' params = list(commodity_name="CORN", 
-#'               year=2012, 
-#'               agg_level_desc = "STATE",
-#'               state_alpha = "WA",
-#'               statisticcat_desc = "YIELD")
-#' nassqs_GET(params)
-#' 
-#' # Equivalent to 'nassqs_record_count(params)'
-#' req <- nassqs_GET(params, api_path = "get_counts")
-#' nassqs_parse(req)
-#' 
-#' # Get the list of allowable values for the parameters 'statisticcat_desc'.
-#' # Equivalent to 'nassqs_param_values("statisticcat_desc")'
-#' req <- nassqs_GET(list(param = "statisticcat_desc"), 
-#'                   api_path="get_param_values")
-#' nassqs_parse(req, as = "list")
-#' }
-nassqs_GET <- function(params, # a named list of queries
-                       key = nassqs_auth(), # API key
-                       api_path=c("api_GET", "get_param_values", "get_counts"), # specific sub API call
-                       base_url="https://quickstats.nass.usda.gov/api/",
-                       url_only = FALSE,
-                       format=c("JSON", "XML", "CSV")) {
-  
-  if(missing(key)) { 
-    key <- tryCatch(nassqs_auth(), error = function(e) e)
[…]
