### 2019-07-24 #298 @potterzot
__Question__: Since both reviewers recommend removing `release_questions()`, where is it recommended that it go? These are helpful (for me) pre-release questions and the function is suggested by the R Packages book ([here](http://r-pkgs.had.co.nz/release.html)), but it doesn't suggest where to put it.

# Response to Reviewers

The reviewers raise some excellent points to consider about usability and organization. The package feels greatly improved by virtue of their comments and suggestions. Thank you again for your time and your invaluable suggestions. `rnassqs` is much cleaner and much improved as a result.

Below I detail some more general thoughts that were raised by the reviewers and my response, and then detail specific response to each reviewer separately.

#### Data size and usability

One reviewer suggestion was to improve the interface to make it more user friendly, i.e. that the package seemed to be built around the needs of the API rather than the needs of the user. To some extent this is a function of the inflexibility of the API itself. A data request to the Quick Stats API returns JSON that, when parsed to a data.frame, results in a data.frame that has 39 columns. Unfortunately there is no way to limit the number of columns returned. The idea of being able to `select` a la `dplyr` is great, but since all 39 columns must be returned, it seems best to leave it to the user to `select` after the call is made.

#### Parameter passing

In response to suggestions about parameters and user experience I've made two changes. The first allows for specifying either a list of parameters in `nassqs` as was the previous case, or specifying each parameter as a separate argument to `nassqs`, as was suggested by @nealrichardson. In addition, I've added links to parameter documentation in the API, and now `nassqs_params()` returns a list of parameters, while `nassqs_params("agg_level_desc")` returns a description of the "agg_level_desc" parameter. I've also updated the vignette to show both methods.

#### Pagination and handling requests larger than 50,000 records

The question of pagination came up repeatedly. It's an interesting one in the context of this API. There is not a direct way to paginate results that the API supports. Typically I end up subsetting by year or geography to make the query small enough. `rnassqs` could try to subset by year or geography automatically, perhaps with a series of rules that first subset by year and if only one year is requested or the request is too large, then also subsetting by a smaller geography. However, there are potential issues here. For example, a state-level request will not necessarily result in the same data as collecting all counties for that state for two reasons:

1. Some data are collected at the state level only
2. County-level data may be suppressed where the state level data is not

Automatically subsetting by year would be doable though. I have added an issue for a future release to do so. In the meantime I have also added information in the error message to suggest how to subset the query. I have also included information on iteration to subset queries in the vignette.

### Specific changes
* Added reviewers to the DESCRIPTION with role = 'rev'
* Fixed spelling errors (sorry)
* Removed `release_questions()` (but see question at the top of this response)
* Added a progress bar as suggested by @nealrichardson

## Response to Reviewer 1

@adamhsparks provided some excellent feedback on issues of usability and potential unnecessary complexity. In particular, this comment was helpful.

> I feel a bit like the functionality for the end-user is a bit overly complicated. This sort of package to me should just return data in a data.frame or list or some other R object that I can easily work with. In some cases I get a server response, in others I can ask for raw JSON or other file formats. I don't really see a good reason for this in R. I guess the functionality could be offered, but I think it needs to be hidden under a few layers for more advanced users that might want to use it. Rather it should just be a simple query for the data that I want and returning it in a format that I can readily use in R without extra steps. Just fetch the data and format it for me in one single step. I think it's possible with this package, but the documentation is a bit convoluted and not clear to me.
There seem to be two major and related issues

While it's true that the package contains `nassqs`, which will just simply query and return a data.frame object without the user having to specify anything, I have reorganized and rewritten the documentation and vignette to emphasize `nassqs` rather than the low-level functions `nassqs_GET` and `nassqs_parse`. I have added documentation and changed the vignette to focus on the ease of use aspect, rather than on building a query using the core functions.

The second issue concerned the organization of functions. I have reorganized functions into files by collective functionality, in an effort to meet the guidelines suggested in [R Packages](http://r-pkgs.had.co.nz/r.html), which states

> While you’re free to arrange functions into files as you wish, the two extremes are bad: don’t put all functions into one file and don’t put each function into its own separate file.

Now functions dealing with the request and parsing of the request are in `request.R`. Authorization functions are in `auth.R`. Helpers are in `helpers.R`. Functions dealing with parameters and parameter values are in `params.R`. Functions that make queries easier are in `wrappers.R`. I think this strikes a good balance, but am certainly open to suggestions to make this clearer if needed.

### Specific Items

#### General package
* Clarified keeping the API key in the `.Renviron` file in the README and vignette
* Added a CITATION file, though it has placeholders until/if the article is accepted by JOSS
* Resolved GIT merge conflicts in paper.md
* contributing and code of conduct text added in CONTRIBUTING.md and CONDUCT.md, as well as in the README

#### Vignette
* Fixed references to `nassqs_fields()` in the vignette, which now refer to `nassqs_param()`
* Fixed error with `nassqs_field_values()` in the vignette
* Fixed inconsistencies with 'PARSE' in the vignette

#### Code
* I did not change the case of `nassqs_GET`, based on style guidelines from the `httr` package: [Best practices for API packages](https://cran.r-project.org/web/packages/httr/vignettes/api-packages.html)
* removed `base_url` as a parameter in `nassqs_GET`

#### Code Style
* Fixed inconsistencies with "=" and "<-"
* All code lines are 80 characters or less except where required (e.g. documentation linking to long urls)

#### Documentation
* Fixed incorrect reference to jsonlite::fromJSON
* Fixed link to utils::read.table
* Added "\code{\link[]{}}" for linking to function documentation (and later removed in favor of markdown syntax).
* I did not change function titles to title case based on the examples from the 'R Packages' [documentation chapter]](http://r-pkgs.had.co.nz/man.html)
* Added an example to `nassqs_params()`
* Fixed example for `nassqs_param_values()`
* Fixed inconsistencies between use of single and double quotes, though note that I use single quotes to refere to keys in a list or object, and double quotes to refer to strings.
* wrapped all examples that make an API call in '\donttest'
* Clarified documentation and examples in 'nassqs_GET' to specify that it is a low-level function
* Clarified documentation in 'nassqs_parse' to better explain what it does and why

## Response to Reviewer 2

@nealrichardson brought up several excellent points about the API and especially about testing and ease of use and focusing on the needs of the user. I feel it is easier to define a list of parameters and submit that as a single argument to `nassqs`, especially for example when iterating over a collection of queries. However, I recognize both needs, and have made it possible to call `nassqs` in either of two ways:

    # First method, a named list of parameters
    params <- list(agg_level_desc = "STATE",
                   state_alpha = c("VA", "WA"))
    nassqs(params)

    # Second method, separate arguments
    nassqs(agg_level_desc = "STATE", state_alpha = c("VA", "WA"))

    # Or without capitalizing
    nassqs(agg_level_desc = "state", state_alpha = c("va", "wa"))

I have expanded the vignette to demonstrate both methods and to emphasize the iteration and pagination of data available by iterating over a list of parameter lists.

Many of @nealrichardson's suggestions involve simplifying the interface, and I think the new function calls are much improved in this regard. These suggests were a real gem. Authorization is simpler, functions have fewer and simpler arguments, and overall ease of use is improved. His suggestion of allowing `year >= 2012` instead of (or in addition to) `year__GE = 2012` is also a good one. I have not implemented it here because I suspect `LIKE` and `NOT LIKE` would be slightly more difficult. I have created an issue to implement this in a future release.

Another concern was that all data is in character format rather than numeric for columns that are numeric. The reason is that the Quick Stats data lists suppressed or unavailable information in a variety of character-based ways. As a result the `Value` field may contain "(D)", "(Z)", or "(S)" rather than numbers. Converting to numeric makes these values `NA`, which loses the specific information about why the data is missing. It is true that it is easy enough to convert to numerical format, but in my opinion keeping this information about why data is missing is important.

The `httptest` package is a huge help and I wish I had known about it when I was asking on twitter about API testing months ago. I've reorganized and updated the tests to use mock API calls, and also with tests in test files that correspond to the function file names in the `R` directory. For example, `test-requests.R` contains tests for functions in `R/request.R`.

### Specific Items

#### General
* Added 'assertthat' to Suggests in DESCRIPTION
* Added stats and utils to Imports in DESCRIPTION
* Added rnassqs.Rcheck and rnassqs_*.tar.gz to .Rbuildignore and .gitignore
* Added case insensitivity
* Converted documentation to markdown format
* Improved documentation generally, especially for parameter definitions
* Added a package web site with `pkgdown`

#### Code
* Moved `nassqs` to top of file since it is the main function
* Fixed `nassqs_fields <- nassqs_params`
* In `nassqs_GET`, removed `url_only` and `format` as function parameters
* In `nassqs_parse`, used `inherits` rather than `class() ==`.
* In `nassqs_parse` RE: "L153: when is not a response object?", at times the API is not working, so in that case this returns the error message directly.
* In `nassqs_parse`, removed unreachable warning
* In `nassqs_parse`, added else case and changed `read.table` to `read.csv`
* In `nassqs` and `nassqs_parse`, change 'raw' to 'text' as an option for the `as` parameter
* In `nassqs_check`, RE: "what do I do if my request is too large? Give some recommendation for how to fix it. If it means that you need to make paginated requests, how would I do that? An even friendlier solution would be for you to handle the pagination for the user so they don't need to know/care about this API constraint."
* In `expand_list`, added @keywords internal and expanded with a description of what it does and why.
* In `nassqs_auth()`, now `nassqs_GET` checks for the environmental variable `NASSQS_TOKEN`, and `nassqs_auth` simply sets that token.
* In `nassqs_parse(), RE: "It was surprising that the one column name that the package code changes gets a rather unfriendly name ("CV (%)"), while the other columns follow a standard, well-behaved format." This name change occurs because of the CSV response, so it is changed there to match the column name in other response types (i.e. JSON and XML)

#### Code Style
* Fixed inline if statements: `if() { }` by removing braces
* Regarding this comment: '@importFrom httr function_name rather than httr::function_name', [R Packages]() generally recommends using the "::" version since it makes clear what package the function is coming from, with the exception that it makes things very slightly slower. However, the vast amount of time in the GET request is due to the API service, so in this case using "::" makes sense to me
* Removed quotes from list names

#### Vignette
* Removed `rnassqs::` from functions
* Removed `.secret` and references to it.

#### Tests
* Removed use of `.secret` for testing
* Reorganized tests to match the file names of the functions they are testing
* Make use of `httptest::with_mock_api`
* Fixed `is.numeric(as.numeric(...))`
