### 2019-07-24 d333105
web site update and minor edits.
