+++ README.Rmd
@@ -86,6 +86,8 @@ Development
 
 This is a package that allows users to access the USDA's National Agricultural Statistics Service (NASS) 'Quick Stats' data through their API. It is fairly low level and does not include a lot of scaffolding or setup. Some things may change, but at this point it is relatively stable.
 
+Visit the package web page here: [https://potterzot.com/rnassqs/](https://potterzot.com/rnassqs/).
+
 ## Installing
 
 Install the package via `devtools` or CRAN:
@@ -129,7 +131,7 @@ Parameters __do not__ need to be capitalized, and also do not need to be in a li
 You can request data for multiple values of the same parameter by using a simple list as follows:
     
     params <- list(commodity_desc = "CORN", year__GE = 2012, state_alpha = c("VA", "WA"))
-    d = nassqs(params=params)
+    d <- nassqs(params=params)
 
 NASS does not allow GET requests that pull more than 50,000 records in one request. The function will inform you if you try to do that. It will also inform you if you've requested a set of parameters for which there are no records.
 
@@ -166,11 +168,11 @@ The NASS API handles other operators by modifying the variable name. The API can
 
 For example, to request corn yields in Virginia and Pennsylvania for all years since 2000, you would use something like:
 
-    params = list(commodity_desc = "CORN", 
+    params <- list(commodity_desc = "CORN", 
                   year__GE = 2000, 
                   state_alpha = c("VA", "PA"), 
                   statisticcat_desc = "YIELD")
-    df = nassqs(params) #returns data as a data frame.
+    df <- nassqs(params) #returns data as a data frame.
 
 See the [vignette](vignettes/rnassqs.html) for more examples and details on usage.
 
