+++ README.Rmd
@@ -174,7 +174,7 @@ For example, to request corn yields in Virginia and Pennsylvania for all years s
                   statisticcat_desc = "YIELD")
     df <- nassqs(params) #returns data as a data frame.
 
-See the [vignette](vignettes/rnassqs.html) for more examples and details on usage.
+See the [vignette](https://potterzot.com/rnassqs/articles/rnassqs.html) for more examples and details on usage.
 
 ## Contributing
 
