### 2019-07-24 0ccafe7
hopefully fix travis API Key Encryption.
