+++ .travis.yml
@@ -10,5 +10,5 @@ r_github_packages:
 after_success:
 - Rscript -e 'library(covr); coveralls()'
 env:
-  matrix:
-  - secure: J/Vyw1RQwTHi5Dqc/nUSgFyfO8SQ2741jdqASRmgDpvTTewn1gJqrX1gzEutqp4E4D/ysnQKj2qg60ZOin1U1RiAUQcvhvd1giJHLmHscxjO/E/kixd8mtYJPVpH6mTl67uq+nmxWeJ+iZceDdFnhpQmUMGzcuAZrsvCoRhPutsxJPSOyrHeeXLzfMNsTsv6GI0Kgy6YZCpxjvwyIJYNwbjXMC1aaPK6iEUSMY6qFwa08FIK87EG24MS+t36URwiMPrIs88DmQXbpIw1HV4h7JW0lv5xKPO2rq3KLc+0y6vqQC0miZK5PCphbz0zehyC02azBB9DtE3+BMXUehq25wCDPmXoXIKWaRB4G1+xYRSHbpw0PAE6zbIdwMQPiYa3SekFtRaIXaRwem6YiuZUdQT2H9BD3lmDJE+W0ZM0bYlDtSIwdJ212I8hnGHSzFPK9sPyQcStMOx1wGWAuZUzHjIEDdFaZduw1Aj2mpLangdrs0IJWm/LqR+HJw9zEDMLGijIgmQ3JG4573gR880TehRKD2Nkf9i9Rillcb0GQ68RnyXc+pELRxlNxTtSSlJRMxbriDspjcpUdunfNiFervYRJHSk8vS5cPHLHbk6xAaWg6EqQlg79Gjt45ry2gwzEbFZXnkRLrXhB9+ZIhXp4db/vrjiPaiRRQ6YYLgwcgI=
+  global:
+    secure: YTwBBmtY7IEiMQonmxk4mq98TbHMvIarC7E1ciYtizhaIfSlYHaYi9RqWkXBl1sRr/b3sTzr7PSUB8iTesmEJXZ8iKYwgUv/adftZUSXiKNd9sJnQvM4L53FHUo7yW1koAe+g3PSg3P6SUlxYj8s8hopWr/in/fJiArto7hFFxQCOv3yZa7Nz4Y7tGT3+tJE4pVAsmrVYAJ0nMdHCsgZNLAChXkFRfNrBkZmoJRRynMDUHuQzsocn14EOrvFerCr0bBmYssxb+GrXaPVU8zf+86Rt8kFJSTJ3uMbn9HEWhurluJqp0SgEFNTT78LAD7UKSBeXz0SRwyQbKmBqN+mG1YD0p90rMMNC7gOojpKz/RO/tr6XhBy6cnyk4zi934D1nFEUfY0NudEyTajrvOfXJhsS09UEUNwWR77cLUdVZhUGe5fFVOBZ+jOjkQKylOO/WwTpN6/96NJ6u4H+XkmS+Wnnv8cmeZGQXKV8/W3DsZFUyWI6lJsLoJx9VrQy5FMX4UKffwhcDKTp0rk1TrdPGSq2onJkyKA801WEAhA9xa+2z56Ov3vK3nstG7Y3PLsjTk5EeXlc9D0yVYkQ1tuSlihiRgvB1gpIIw7l+Ha0uTw735K3Kr3Q1ux03n2ih5SK/Ey8c7YfMjhxT7OpWwk5pItKL2nlB13igJK7qIIc78=
