+++ README.Rmd
@@ -7,7 +7,7 @@ output:
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
 ![rnassqs icon](profile.png)
-(<p style="font-size: 0.9rem;font-style: italic;"><a href="https://www.flickr.com/photos/53018729@N00/2669034542">"wheat"</a><span>by <a href="https://www.flickr.com/photos/53018729@N00">ani!</a></span> is licensed under <a href="https://creativecommons.org/licenses/CC0/1.0/?ref=ccsearch&atype=html" style="margin-right: 5px;">CC CC0 1.0</a><a href="https://creativecommons.org/licenses/CC0/1.0/?ref=ccsearch&atype=html" target="_blank" rel="noopener noreferrer" style="display: inline-block;white-space: none;opacity: .7;margin-top: 2px;margin-left: 3px;height: 22px !important;"><img style="height: inherit;margin-right: 3px;display: inline-block;" src="https://ccsearch.creativecommons.org/static/img/cc_icon.svg" /><img style="height: inherit;margin-right: 3px;display: inline-block;" src="https://ccsearch.creativecommons.org/static/img/cc-cc0_icon.svg" /></a></p>)
+(Wheat image from: <p style="font-size: 0.9rem;font-style: italic;"><a href="https://www.flickr.com/photos/53018729@N00/2669034542">"wheat"</a><span>by <a href="https://www.flickr.com/photos/53018729@N00">ani!</a></span> is licensed under <a href="https://creativecommons.org/licenses/CC0/1.0/?ref=ccsearch&atype=html" style="margin-right: 5px;">CC CC0 1.0</a><a href="https://creativecommons.org/licenses/CC0/1.0/?ref=ccsearch&atype=html" target="_blank" rel="noopener noreferrer" style="display: inline-block;white-space: none;opacity: .7;margin-top: 2px;margin-left: 3px;height: 22px !important;"><img style="height: inherit;margin-right: 3px;display: inline-block;"</p>)
 
 <table class="table">
 <thead>
