### 2019-07-24 60c2ff4
small fixes to readme image.
