### 2019-07-24 1645e49
Update NEWS and README to reflect changes.
