### 2019-07-24 d57d356
update vignette to fetch all years since 1997 in iterating example.
