### 2019-07-24 0ce56ed
update README for hexsticker.
