+++ README.Rmd
@@ -6,12 +6,11 @@ output:
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-![rnassqs icon](profile.png)
-(Wheat image from [here](https://www.flickr.com/photos/53018729@N00/2669034542))
 
 <table class="table">
 <thead>
 <tr class="header">
+<th></th>
 <th align="left">
 Usage
 </th>
@@ -26,6 +25,10 @@ Development
 <tbody>
 
 <tr class="odd">
+<td rowspan="5">
+  ![rnassqs icon](inst/figures/rnassqs.png)
+  (Wheat image from [here](https://www.flickr.com/photos/53018729@N00/2669034542))
+</td>
 <td align="left">
 <a href="http://choosealicense.com/licenses/mit/"><img src="https://img.shields.io/github/license/mashape/apistatus.svg" alt="License"></a>
 </td>
