### 2019-07-24 bc7e7c8
make hexsticker smaller.
