### 2019-07-24 4a929a2
change paths in hexsticker creation to be run from project root.
