### 2019-07-24 4ef56eb
fix minor error in README usage.
