+++ README.Rmd
@@ -140,7 +140,7 @@ Parameters __do not__ need to be capitalized, and also do not need to be in a li
 You can request data for multiple values of the same parameter by using a simple list as follows:
     
     params <- list(commodity_desc = "CORN", year__GE = 2012, state_alpha = c("VA", "WA"))
-    d <- nassqs(params=params)
+    d <- nassqs(params)
 
 NASS does not allow GET requests that pull more than 50,000 records in one request. The function will inform you if you try to do that. It will also inform you if you've requested a set of parameters for which there are no records.
 
