### 2019-07-24 917b2c6
fix minor erro in README about key handling.
