### 2019-07-25 #298 @adamhsparks
This looks much improved as I've glanced over it. Thanks for thoughtfully responding to our reviews and comments. Thanks for explaining any reasons why my suggestions weren't followed, I have no objections to any of them. Some of my suggestions have been based on CRAN's, umm, erratic(?) enforcement of rules from time-to-time, so ignoring some of my suggestions are probably fine as I'm not sure that I use title case in all my documentation everywhere but think I was pulled up on it once before.

I think that the organisation of the functions is much more clear now and agree with Hadley on not all in one and not one only per .R file.

Regarding the question on `release_questions()`, `devtools::release()` actually asks most of those and more when you use it, I'd suggest using that rather than including questions for yourself in the package NAMESPACE.

For `expand_list()` I'd use a `@noRd` tag since it does not need to be exposed to the end-user, that I can tell? Documentation is good, I do that for my internal functions so I know what they do, but it shouldn't clutter the user's experience having it documented unless it's used somewhere that I'm missing where an end-user actually calls it?

There's no need for the CITATION file to be incomplete as you've suggested. It should have two entries after acceptance to JOSS. One should just be for the package, the current version number and year it was released that will automatically update with new releases, which you can set up now. The second is the JOSS paper citation that will never change. The example I provided shows this.

I'm curious, how is it different than `usdarnass`, which you have mentioned in the README now? This isn't detailed in the original submission.
