### 2019-07-25 #298 @potterzot
@adamhsparks thank you. I've updated the CITATION file and also added @noRd to `expand_list()`.

Regarding `usdarnass`, I added the reference to the README after I found out about it, which was after I had submitted for rOpenSci review. I'm fairly sure `rnassqs` was developed first, since my first git commit was in June 2015, while theirs was November 2018, and `rnassqs` was published on CRAN on May 03 while `usdarnass` was published on CRAN on June 21. I think they were actually developed unaware of each other. If you have any thoughts or suggestions about a course of action I'd be all ears. It seems we could basically continue to develop in parallel or we could merge packages. I haven't reached out to the authors other than make a suggestion on an issue to let them know they could allow for multiple options as I write below.

The differences are small as far as I can tell:

* `rnassqs` makes all of the query parameters available, while `usdarnass` only allows a subset.
* `rnassqs` allows multiple options like `state_alpha = c("VA", "WA")`, but after I commented on an issue on `usdarnass` to say it was possible to do that, `usdarnass` does that as well.
* Before this review, `rnassqs` took a list of parameters like `nassqs(params)`, while `usdarnass` takes parameters like `nass_data(state_alpha = "VA", agg_level_desc = "state")`. Now as per suggestion from @nealrichardson `rnassqs` works either way.
