### 2019-07-25 #298 @nealrichardson
I'll just briefly comment that this all sounds good in principle and I look forward to re-reviewing in detail, though I won't be able to get to that until early next week.
