### 2019-07-25 ce1af80
add Date to DESCRIPTION and fix CITATION.
