+++ DESCRIPTION
@@ -2,6 +2,7 @@ Package: rnassqs
 Type: Package
 Title: Access the NASS 'Quick Stats' API
 Version: 0.4.0.9002
+Date: 2019-07-25 
 Authors@R: c(
   person('Nicholas', 'Potter', email='potter.nicholas@gmail.com', role = c('aut', 'cre')),
   person('Jonathan', 'Adams', email='', role = c('ctb')),
