### 2019-08-01 #298 @lmullen
Thanks for the detailed changes and response, @potterzot.

@adamhsparks and @nealrichardson Thanks for looking over the changes. Could you please complete your re-review and either request additional changes or vote to approve the package by August 8?
