### 2019-08-02 #298 @adamhsparks
Will do!

--
Dr Adam H. Sparks
Associate Professor of Field Crops Pathology  |  Centre for Crop Health
Institute for Life Sciences and the Environment (ILSE)  |  Research and Innovation Division
University of Southern Queensland | Toowoomba, Queensland | 4350 | Australia

Phone: +61746311948
Mobile: +61415489422
Mobile: +61415489422
