### 2019-08-02 4f70efb
Improve test wrappers and skip another test that requires auth
