+++ tests/testthat/test-wrappers.R
@@ -11,65 +11,39 @@ params = list(
 
 ### Test API URLs with mock APIs ----
 
-# First set the API KEY to a static value
-key <- Sys.getenv("NASSQS_TOKEN")
-nassqs_auth(key = "API_KEY")
-
-# Tests
 with_mock_api({
   test_that("nassqs_record_count forms a correct URL", {
     expect_GET(nassqs_record_count(params),
                  "https://quickstats.nass.usda.gov/api/get_counts?key=API_KEY&commodity_desc=CORN&year=2012&statisticcat_desc=YIELD&agg_level_desc=STATE&state_alpha=VA&state_alpha=WA&format=JSON")
   })
-})
 
-with_mock_api({
   test_that("nassqs_yield returns a correct URL", {
     expect_GET(nassqs_yields(params),
                "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&statisticcat_desc=YIELD&agg_level_desc=STATE&state_alpha=VA&state_alpha=WA&format=JSON")
   })
-})
 
-with_mock_api({
   test_that("nassqs_yield returns a correct URL even if statisticcat_desc parameter is set to a different value", {
     params[['statisticcat_desc']] <- "AREA HARVESTED"
     expect_GET(nassqs_yields(params),
                "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&statisticcat_desc=YIELD&agg_level_desc=STATE&state_alpha=VA&state_alpha=WA&format=JSON")
   })
-})
 
-with_mock_api({
   test_that("nassqs_area returns a correct URL", {
     expect_GET(nassqs_acres(params, area = "AREA BEARING"),
                "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&statisticcat_desc=AREA%20BEARING&agg_level_desc=STATE&state_alpha=VA&state_alpha=WA&unit_desc=ACRES&format=JSON")
   })
-})
 
-with_mock_api({
   test_that("nassqs_area returns a correct URL when multiple areas are specified", {
     expect_GET(nassqs_acres(params, area = c("AREA HARVESTED", "AREA PLANTED")),
                  "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&statisticcat_desc=AREA%20HARVESTED&statisticcat_desc=AREA%20PLANTED&agg_level_desc=STATE&state_alpha=VA&state_alpha=WA&unit_desc=ACRES&format=JSON")
   })
 })
 
-# Set the key after finishing mock tests
-nassqs_auth(key = key)
-
 
 ### API tests that require the API ----
-test_that("nassqs_record_count returns a numeric", {
-  skip_if_no_auth()
-  skip_on_cran()
-  
-  v = nassqs_record_count(params)
-  expect_equal(is.numeric(v$count), TRUE)
+with_authentication({
+  test_that("nassqs_record_count returns a numeric", {
+    v = nassqs_record_count(params)
+    expect_equal(is.numeric(v$count), TRUE)
+  })
 })
-
-
-
-
-
-
-
-
-
