+++ tests/testthat/test-request.R
@@ -12,115 +12,88 @@ params = list(
 
 ### Test API URLs with mock APIs ----
 
-# First set the API KEY to a static value
-key <- Sys.getenv("NASSQS_TOKEN")
-nassqs_auth(key = "API_KEY")
-
-
 with_mock_api({
   test_that("nassqs forms a correct URL", {
     expected_url <- "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&agg_level_desc=STATE&statisticcat_desc=AREA%20HARVESTED&domaincat_desc=NOT%20SPECIFIED&state_alpha=VA&format=JSON"
     expect_GET(nassqs(params), url = expected_url)
   })
-})
 
-with_mock_api({
   test_that("nassqs_GET() forms a correct URL", {
     expected_url <- "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&agg_level_desc=STATE&statisticcat_desc=AREA%20HARVESTED&domaincat_desc=NOT%20SPECIFIED&state_alpha=VA&format=JSON"
     expect_GET(nassqs_GET(params), url = expected_url)
   })
 })
 
-# Set the key after finishing mock tests
-nassqs_auth(key = key)
-
 
 ### API tests that make real API calls if local and have an API key ----
-test_that("nassqs_GET successfully makes a request to the Quick Stats API", {
-  skip_if_no_auth()
-  skip_on_cran()
-  
-  req <- nassqs_GET(params)
-  expect_is(req, "response")
-})
+with_authentication({
+  test_that("nassqs_GET successfully makes a request to the Quick Stats API", {
+    req <- nassqs_GET(params)
+    expect_is(req, "response")
+  })
 
-test_that("nassqs_GET with api_path = 'api_GET' successfully makes a request to the Quick Stats API", {
-  skip_if_no_auth()
-  skip_on_cran()
-  
-  req <- nassqs_GET(params, api_path = 'api_GET')
-  expect_is(req, "response")
-})
+  test_that("nassqs_GET with api_path = 'api_GET' successfully makes a request to the Quick Stats API", {
+    req <- nassqs_GET(params, api_path = 'api_GET')
+    expect_is(req, "response")
+  })
 
-test_that("nassqs_GET with api_path = 'get_param_values' successfully makes a request to the Quick Stats API", {
-  skip_if_no_auth()
-  skip_on_cran()
-  
-  req <- nassqs_GET(param = "source_desc", api_path = "get_param_values")
-  expect_is(req, "response")
-  
-  values <- nassqs_parse(req)
-  expect_is(values, "list")
-})
+  test_that("nassqs_GET with api_path = 'get_param_values' successfully makes a request to the Quick Stats API", {
+    req <- nassqs_GET(param = "source_desc", api_path = "get_param_values")
+    expect_is(req, "response")
 
-test_that("nassqs_GET with api_path = 'get_counts' successfully makes a request to the Quick Stats API", {
-  skip_if_no_auth()
-  skip_on_cran()
-  
-  req <- nassqs_GET(params, api_path = "get_counts")
-  expect_is(req, "response")
-  
-  n <- nassqs_parse(req)
-  expect_is(n$count, "integer")
-})
+    values <- nassqs_parse(req)
+    expect_is(values, "list")
+  })
 
-test_that("nassqs_GET() works with lower case values", {
-  lower_params = list(
-    commodity_desc = "corn",
-    year = "2012",
-    agg_level_desc = "state",
-    statisticcat_desc = "area harvested",
[…]
