### 2019-08-02 3cae867
Add example of a test for error response from API
