+++ tests/testthat/test-request.R
@@ -22,6 +22,15 @@ with_mock_api({
     expected_url <- "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&agg_level_desc=STATE&statisticcat_desc=AREA%20HARVESTED&domaincat_desc=NOT%20SPECIFIED&state_alpha=VA&format=JSON"
     expect_GET(nassqs_GET(params), url = expected_url)
   })
+
+  test_that("Too-large request error is handled", {
+    p2 <- params
+    p2$year <- 2013
+    expect_error(
+      nassqs(p2),
+      "Request was too large. NASS requires that an API call returns a max"
+    )
+  })
 })
 
 
