### 2019-08-02 cfa15d5
Move lowercase test to work without auth; refactor tolower/toupper
