+++ R/request.R
@@ -1,16 +1,16 @@
 #' Get data and return a data frame
 #'
 #' The primary function in the `rnassqs` package, `nassqs` makes a HTTP GET
-#' request to the USDA-NASS Quick Stats API and returns the data parsed as a 
+#' request to the USDA-NASS Quick Stats API and returns the data parsed as a
 #' data.frame, plain text, or list. Various other functions make use of `nassqs`
-#' to make specific queries. For a data request the Quick Stats API returns 
-#' JSON that when parsed to a data.frame contains 39 columns and a varying 
+#' to make specific queries. For a data request the Quick Stats API returns
+#' JSON that when parsed to a data.frame contains 39 columns and a varying
 #' number of rows depending on the query. Unfortunately there is not a way to
 #' restrict the number of columns.
 #'
 #' @export
 #'
-#' @param ... either a named list of parameters or a series of parameters to 
+#' @param ... either a named list of parameters or a series of parameters to
 #'   form the query
 #' @param as whether to return a data.frame, list, or text string
 #'   [nassqs_GET()]
@@ -18,15 +18,15 @@
 #' @seealso [nassqs_GET()], [nassqs_yields()], [nassqs_acres()]
 #' @examples \donttest{
 #'   # Get corn yields in Virginia in 2012
-#'   params <- list(commodity_name = "CORN", 
-#'                  year = 2012, 
+#'   params <- list(commodity_name = "CORN",
+#'                  year = 2012,
 #'                  agg_level_desc = "COUNTY",
-#'                  state_alpha = "VA", 
+#'                  state_alpha = "VA",
 #'                  statisticcat_desc = "YIELD")
 #'   yields <- nassqs(params)
 #'   head(yields)
 #' }
-nassqs <- function(..., 
+nassqs <- function(...,
                    as = c("data.frame", "text", "list")) {
   as = match.arg(as)
   req <- nassqs_GET(..., api_path = "api_GET")
@@ -35,55 +35,55 @@ nassqs <- function(...,
 
 
 #' Issue a GET request to the NASS 'Quick Stats' API
-#' 
-#' This is the workhorse of the package that provides the core request 
-#' functionality to the NASS 'Quick Stats' API: 
+#'
+#' This is the workhorse of the package that provides the core request
+#' functionality to the NASS 'Quick Stats' API:
 #' [https://quickstats.nass.usda.gov/api](https://quickstats.nass.usda.gov/api).
-#' In most cases [nassqs()] or other high-level functions should be used. 
-#' `nassqs_GET` uses [httr::GET()] to make a HTTP GET request, which returns a 
+#' In most cases [nassqs()] or other high-level functions should be used.
+#' `nassqs_GET` uses [httr::GET()] to make a HTTP GET request, which returns a
 #' request object which must then be parsed to a data.frame, list, or other `R`
-#' object. Higher-level functions will do that parsing automatically. However, 
-#' if you need access to the request object directly, `nassqs_GET` provides 
+#' object. Higher-level functions will do that parsing automatically. However,
+#' if you need access to the request object directly, `nassqs_GET` provides
 #' that.
 #'
 #' @export
-#' 
-#' @param ... either a named list of parameters or a series of parameters to 
+#'
+#' @param ... either a named list of parameters or a series of parameters to
 #'   use in the query
 #' @param api_path the API path that determines the type of request being made
 #' @return a [httr::GET()] response object
 #' @examples \donttest{
 #'   # Yields for corn in 2012 in Washington
-#'   params <- list(commodity_name = "CORN", 
-#'                  year = 2012, 
+#'   params <- list(commodity_name = "CORN",
+#'                  year = 2012,
 #'                  agg_level_desc = "STATE",
 #'                  state_alpha = "WA",
[…]
