### 2019-08-02 1a653d3
Test a couple other conditions
