### 2019-08-02 1e992c1
Remove skip_if_no_auth, no longer used
