### 2019-08-02 #298 @nealrichardson
Nice work. This looks much improved. I found a few sylistic issues again, and I had some suggestions for how to improve the testing, but rather than write them here, I've made a [pull request](https://github.com/potterzot/rnassqs/pull/15) with them for you to review/merge. The other reason I implemented these suggestions myself was that I got a test failure locally because one of the tests required auth but did not have the appropriate `skip_if_no_auth()` check, so I was already in the code to debug that. There was also a `R CMD check` issue I encountered because the .Rbuildignore still wasn't correctly excluding previously built tarballs. All fixed by that PR.

Test coverage could be better, though my PR bumps it up to 91%. Happy to advise on covering the conditions that currently are missed that if you want, though I won't withhold approval based on not reaching 100% line coverate.

One last followup point:

> In nassqs_parse RE: "L153: when is not a response object?", at times the API is not working, so in that case this returns the error message directly.

[`httr::GET()`](https://httr.r-lib.org/reference/GET.html) will either return a response object (potentially with an error status, which you will have already have handled before getting here because you pass through the `nassqs_check()` function) or `GET()` will itself error (like if your internet is down). I don't think it's possible for it to return anything different.
