### 2019-08-08 #298 @adamhsparks
I have a few last minor points (nitpicks?) that if changed will improve the package. Overall it's greatly improved and I like how it works. Congrats!

1. * @lmullen already noted this much earlier in the process. Please remove the "Date" field from the DESCRIPTION file. CRAN will automatically assign this and it's prone to ending up being out of date if you rely upon updating it manually.

2. You don't need a `paste()` in a `stop()`, _e.g._,
`stop(paste0("Your query parameters include 'format' as ", format,
			" but it should be one of 'json', 'xml', or 'csv'."))`.
It should be written as:
`stop("Your query parameters include 'format' as ", format,
	 " but it should be one of 'json', 'xml', or 'csv'.")`
There are several instances of this that I noted.

3. * "inst/examples/example_parameters.R" has an incomplete final line. Add a line return to the file at the end of the file to fix this.

4. The documentation for `nasqss_GET()` is inconsistent in how it references functions in the description. Some of the other R functions discussed in that paragraph use the `function()` convention while referring to `nasqss_GET()` is only as `nassqss_GET` minus the `()`. As a user I find it more clear if the `()` is used in documentation to indicate a function, not a parameter is being discussed. *Note* I didn't check all documentation, I just noticed this here.

5. * The documentation example for `nassqss_param_values` has "YIElD" not "YIELD" in the comment section, is this correct?

6. * `nassqss_parse()` documentation Description field is missing a "'" prior to (Z).

7. * I'm not sure that _here_ needs to be listed in the DESCRIPTION Suggests field. It's only used in data-raw as far as I can tell? If so, that folder is not included in the R package so shouldn't need to be specified here.

8. * In the data-raw/get_test_data.R file, it might be good to set the `version` to `2` for maximum compatibility in the near term with versions of R from 1.4.0 to current. If it's `NULL` it will default to version 3.

9. The README code could be formatted a bit more nicely too using proper RMarkdown chunks, _e.g._,
    # Via devtools
    library(devtools)
    install_github('potterzot/rnassqs')

    # Via CRAN
    install.packages("rnassqs")```
```

10. The README may not need to be a .Rmd? I can't see that you have any executed R code so you could simplify and just use a .md file.

11. Consider using `codemetar::write_codemeta()` to create and update a .json metadata file for the package?

Once these are addressed (at your discretion for most of them) I'm happy to recommend accepting. I've added a "*" after the number and before the comment for the items that I think must be fixed. Those without are at your discretion.
