### 2019-08-12 #298 @potterzot
@adamhsparks Thank you for the incredible detail in this! Much appreciated. I've made all of the changes you suggest except for this one, which I'm unclear on:

> 8.* In the data-raw/get_test_data.R file, it might be good to set the version to 2 for maximum compatibility in the near term with versions of R from 1.4.0 to current. If it's NULL it will default to version 3.

What do you mean by setting the version? Do you mean setting the `R` version in DESCRIPTION?
