### 2019-08-12 19d8acd
Make changes to address round 2 revisions from @adamhsparks
- remove the 'Date' field in DESCRIPTION
- remove 'paste0' in `stop()` calls
- add an empty line to `/inst/examples/example_parameters.R`
- change `nassqs_GET()` documentation to be consistent
- fix miscapitalizing of 'YIElD'
- add proper R evaluation to README.Rmd
- update codemeta::write_codemeta()
