+++ inst/examples/example_parameters.R
@@ -81,3 +81,4 @@ ag_sales = list(sector_desc = "ECONOMICS",
                 domain_desc = "TOTAL",
                 domaincat_desc = "NOT SPECIFIED",
                 year__GE = 1997)
+
