+++ README.Rmd
@@ -101,19 +101,22 @@ For docs and code examples, visit the package web page here: [https://potterzot.
 
 Install the package via `devtools` or CRAN:
 
+```{r eval=FALSE}
     # Via devtools
     library(devtools)
     install_github('potterzot/rnassqs')
     
     # Via CRAN
     install.packages("rnassqs")
-
+```
 
 ## API Key
 
 To use the NASS Quick Stats API you need an [API key](http://quickstats.nass.usda.gov/api). The API key should in general not be included in scripts. One way of making the key available without defining it in a script is by setting it in your `.Renviron` file, which is usually located in your home directory. If you are an `rstudio` user, you can use `usethis::edit_r_environ()` to open your `.Renviron` file and add a line that looks like:
 
+```{r eval=FALSE}
     NASSQS_TOKEN="<your api key here>"
+```
 
 Alternatively, you can set it explicitly in the console with `nassqs_auth(key = <your api key>)`. This will set the environmental variable NASSQS_TOKEN, which is used to access the API. You can also set this directly with `Sys.setenv("NASSQS_TOKEN" = <your api key>)`. 
 
@@ -124,6 +127,7 @@ See the examples in [inst/examples](inst/examples) for quick recipes to download
 The primary function is `nassqs()`, with which you can make any query of variables. 
 For example, to mirror the request that is on the [NASS API documentation](http://quickstats.nass.usda.gov/api), you can use:
 
+```{r eval=FALSE}
     library(nassqs)
     
     # You must set your api key before requesting data
@@ -132,20 +136,26 @@ For example, to mirror the request that is on the [NASS API documentation](http:
     # Parameters to query on and data call
     params <- list(commodity_desc = "CORN", year__GE = 2012, state_alpha = "VA")
     d <- nassqs(params)
+```
 
 Parameters __do not__ need to be capitalized, and also do not need to be in a list format. The following works just as well:
 
+```{r eval=FALSE}
     d <- nassqs(commodity_desc = "corn", year__GE = 2012, state_alpha = "va")
+```
 
 You can request data for multiple values of the same parameter by using a simple list as follows:
     
+```{r eval=FALSE}
     params <- list(commodity_desc = "CORN", year__GE = 2012, state_alpha = c("VA", "WA"))
     d <- nassqs(params)
+```
 
 NASS does not allow GET requests that pull more than 50,000 records in one request. The function will inform you if you try to do that. It will also inform you if you've requested a set of parameters for which there are no records.
 
 Other useful functions include:
 
+```{r eval=FALSE}
     # returns a set of unnique values for the parameter "STATISTICCAT_DESC"
     nassqs_param_values("statisticcat_desc")
     
@@ -162,6 +172,7 @@ Other useful functions include:
     
     # Specifies just "AREA HARVESTED" values of statisticcat_desc
     nassqs_acres(params, area = "AREA HARVESTED")
+```
 
 
 ### Handling inequalities and operators other than "="
@@ -177,11 +188,13 @@ The NASS API handles other operators by modifying the variable name. The API can
 
 For example, to request corn yields in Virginia and Pennsylvania for all years since 2000, you would use something like:
 
+```{r eval=FALSE}
     params <- list(commodity_desc = "CORN", 
                   year__GE = 2000, 
                   state_alpha = c("VA", "PA"), 
                   statisticcat_desc = "YIELD")
     df <- nassqs(params) #returns data as a data frame.
+```
 
 See the [vignette](https://potterzot.com/rnassqs/articles/rnassqs.html) for more examples and details on usage.
 
