+++ DESCRIPTION
@@ -1,8 +1,7 @@
 Package: rnassqs
 Type: Package
 Title: Access the NASS 'Quick Stats' API
-Version: 0.4.0.9002
-Date: 2019-07-25 
+Version: 0.4.0.9003
 Authors@R: c(
   person('Nicholas', 'Potter', email='potter.nicholas@gmail.com', role = c('aut', 'cre')),
   person('Jonathan', 'Adams', email='', role = c('ctb')),
@@ -28,7 +27,6 @@ Imports:
   utils
 Suggests: 
   assertthat,
-  here,
   httptest,
   knitr,
   rmarkdown,

[…]
