### 2019-08-12 ed48ee6
Update documentation and fix issues with tests
Two tests in `/tests/testthat/test-request.R` were within the
`with_mock_api({})` block but required actual API calls.
