+++ tests/testthat/quickstats.nass.usda.gov/api/api_GET-a17d57.R
@@ -1,8 +1,0 @@
-structure(
-  list(
-    url = "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2102&agg_level_desc=STATE&statisticcat_desc=AREA%20HARVESTED&domaincat_desc=NOT%20SPECIFIED&state_alpha=VA&format=JSON",
-    status_code = 404,
-    content = charToRaw('{"message": "Not Found"}')
-  ),
-  class = "response"
-)
