+++ tests/testthat/quickstats.nass.usda.gov/api/api_GET-559ed7.R
@@ -1,7 +1,0 @@
-structure(
-  list(
-    url = "https://quickstats.nass.usda.gov/api/api_GET?key=API_KEY&commodity_desc=CORN&year=2012&agg_level_desc=STATE&statisticcat_desc=AREA%20HARVESTED&domaincat_desc=NOT%20SPECIFIED&state_alpha=VA&format=JSON",
-    status_code = 413
-  ),
-  class = "response"
-)
