### 2019-08-12 3d28070
Add @nealrichardson as a contributor.
