### 2019-08-12 #298 @potterzot
@nealrichardson I've reviewed and merged your PR, thanks! There were two tests for error handling that were failing:

- "Too-large request error is handled"
- "Other server error is handled"

Because they were within the `with_mock_api()` block they were returning a `GET` object instead of the error. I moved these to the authorization block and they work.

Regarding `.Rbuildignore` excluding tarballs, I had included that at some point long ago, but removed it for a reason that I don't remember. Thank you for adding it.
