### 2019-08-12 #298 @potterzot
@nealrichardson PS if I have your okay I've also added you as a contributor in DESCRIPTION.
