### 2019-08-12 #298 @nealrichardson
Where did you see the failure? [The PR merge commit passed on Travis](https://travis-ci.org/potterzot/rnassqs/builds/570904679). They don't require auth because they use the mock responses I added here: https://github.com/potterzot/rnassqs/pull/15/files#diff-7c5a672790a8227968bfd57c3a71faa0 Did you possibly make other changes that altered the querystring in the request? That would change the request URL and thus change the mock file path it was looking for. If so, you can rename those mock files to match and they'll be fine.

Sure, happy to be listed at contributor.
