### 2019-08-12 a70fe5e
remove unnecessary check if a response.
