### 2019-08-12 #298 @potterzot
@nealrichardson Hmm, I checked out your commit again and am having no trouble. I must have changed something that started giving me those errors. I returned those files to their original state in a new commit. Also removed the response check based on your note about `nassqs_GET()` always returning a response object.
