### 2019-08-12 #298 @adamhsparks
Hi @potterzot,
The RDS version of the file is 1, 2 or (default) 3. This changed with R 3.6 to "3". So many users may not have the ability to read a version 3 RDS file yet if they've not upgraded to R >= 3.6.

See `?saveRDS` for more on `version`
