+++ data-raw/get_test_data.R
@@ -3,13 +3,6 @@
 # Generate response data for tests
 library(rnassqs)
 
-key_file <- here::here(".secret")
-if(file.exists(key_file)) {
-  con <- file(key_file)
-  key <- readLines(con)
-  close(con)
-}  
-
 params = list(
   commodity_desc = "CORN",
   year = "2012",
@@ -19,15 +12,7 @@ params = list(
 )
 
 ### JSON
-req <- nassqs_GET(params, key = key, format = "JSON")
-saveRDS(req, file = here::here("tests/data/request_json.rds"))
-
-### XML
-req <- nassqs_GET(params, key = key, format = "XML")
-saveRDS(req, file = here::here("tests/data/request_xml.rds"))
-
-### CSV
-req <- nassqs_GET(params, key = key, format = "CSV")
-saveRDS(req, file = here::here("tests/data/request_csv.rds"))
+req <- nassqs_GET(params)
+saveRDS(req, file = here::here("tests/testthat/testdata/request_json.rds"), version = "2")
 
 
