### 2019-08-13 #298 @adamhsparks
@lmullen, I've updated my initial review with the suggestion to accept, ticked the rest of the boxes and updated time spent reviewing.
