### 2019-08-15 #298 @lmullen
@adamhsparks Great, thanks so much.

@nealrichardson Is there anything else outstanding from your perspective?
