### 2019-08-15 #298 @lmullen
Approved! Thanks @potterzot for submitting and making all the requested changes. And thanks, @adamhsparks and @nealrichardson for especially thorough reviews. Much appreciated.

@potterzot here are some to-dos to complete the onboarding process.

- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do. Once you do I'll give you admin access.

"```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)```"

Since you are also publishing to JOSS, we need you to do the following.
- [ ] Submit to JOSS using the Zenodo DOI.  http://joss.theoj.org/papers/new When the paper shows up at JOSS, then add the following comment to the submission thread. `This submission has been accepted to rOpenSci. The review thread can be found at <URL TO THIS REVIEW>.`

You can also release a new version to CRAN.

Welcome aboard! We'd love to host a blog post about your package - either a [short introduction to it with one example](https://ropensci.org/tech-notes/) or [a longer post with some narrative about its development or something you learned, and an example of its use](https://ropensci.org/blog/). If you are interested, review [the instructions](https://github.com/ropensci/roweb2#contributing-a-blog-post), and tag @stefaniebutland in your reply. She will get in touch about timing and can answer any questions.

We've started putting together a gitbook with our best practice and tips, [this chapter](https://ropensci.github.io/dev_guide/collaboration.html) starts the 3d section that's about guidance for after onboarding. Please tell us what could be improved, the corresponding repo is [here](https://github.com/ropensci/dev_guide).
