### 2019-09-10 #298 @stefaniebutland
@potterzot Sounds good! Please submit a draft when you're ready and we can select a publication date at that time.
> why I started developing the package and how it's been helpful in my research

If you include a "cool" example (that's not shown elsewhere) this is especially valuable as a way for readers to see how they might use the package.

> what I've learned in the process

Always good to share this. Try to choose a couple of key points.

Thanks!
