### 2019-11-08 #298 @potterzot
@stefaniebutland I couldn't make the Oct 29 deadline but have a working draft now, do you have a good future date that would work? The draft is basically done but I can change the template date and file names to match the anticipated date.

Also, I'm not sure where to put images. The template links I can change, but I don't see the corresponding `img/blog-images directory in the `roweb2` repository.
