### 2019-11-08 #298 @potterzot
@stefaniebutland nevermind on the second part, I figured out the images. It helps if I read the instructions in full!
